+++ playground/supabase-auth/app/api/test/route.ts
@@ -26,34 +26,32 @@ export const GET = async () => {
   // if there are no current subscriptions, create a new one
   // otherwise, use the first one
   let currentSubscriptionId: string | undefined;
-  if (!currentSubscriptions || currentSubscriptions.length === 0) {
+  if (!currentSubscriptions || currentSubscriptions.length === 0 || true) {
     const { subscription: newSubscription } =
       await flowgladServer.createSubscription({
         priceId,
         trialEnd: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30).getTime()
       });
     currentSubscriptionId = newSubscription.id;
-  } else {
-    currentSubscriptionId = currentSubscriptions[0].id;
   }
   if (!currentSubscriptionId) {
     throw new Error('No current subscription id found');
   }
   // create a usage event
-  const usageEvent = await flowgladServer.createUsageEvent({
-    usageMeterId,
-    transactionId: currentSubscriptionId,
-    subscriptionId: currentSubscriptionId,
-    priceId,
-    amount: 1,
-    properties: {
-      githubOrganization: 'flowglad',
-      user: 'lol'
-    }
-  });
+  // const usageEvent = await flowgladServer.createUsageEvent({
+  //   usageMeterId,
+  //   transactionId: currentSubscriptionId,
+  //   subscriptionId: currentSubscriptionId,
+  //   priceId,
+  //   amount: 1,
+  //   properties: {
+  //     githubOrganization: 'flowglad',
+  //     user: 'lol'
+  //   }
+  // });
 
   return NextResponse.json({
-    message: 'Hello, world!',
-    usageEvent
+    message: 'Hello, world!'
+    // usageEvent
   });
 };