+++ packages/express/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @flowglad/express
 
+## 0.10.11
+
+### Patch Changes
+
+- Fix FlowgladThemeProvider styles
+- Updated dependencies
+  - @flowglad/react@0.10.11
+  - @flowglad/server@0.10.11
+  - @flowglad/shared@0.10.11
+
 ## 0.10.10
 
 ### Patch Changes